requirements.txt
fastapi
uvicorn
pydantic
torch
transformers
omegaconf
pyyaml
pinggy
requests
sentencepiece
python-multipart